+++ src/Client.ts
@@ -5,6 +5,7 @@ import Codes from './Resources/Codes/Codes';
 import Subscriptions from './Resources/Subscriptions';
 import { loadWaccResources, WaccResources } from './Resources/Wacc';
 import { loadGuestsResources, GuestsResources } from './Resources/Guests';
+import { loadGlobalResources, GlobalResources } from './Resources/Global';
 import Users from './Resources/Users';
 import Gift from './Resources/Gift/Gift';
 import Products from './Resources/Products';
@@ -18,6 +19,7 @@ export default class Client {
     public guests!: GuestsResources;
     public users!: Users;
     public gift!: Gift;
+    public global!: GlobalResources;
 
     protected httpClient: HttpClient;
     protected options: ClientAllOptions = {
@@ -49,6 +51,7 @@ export default class Client {
         this.guests = loadGuestsResources(this);
         this.users = new Users(this);
         this.gift = new Gift(this);
+        this.global = loadGlobalResources(this);
     }
 
     getBasePath(): string {