+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🚀 Features
+- Add global/contact endpoint for submitting contact form ([#187](https://github.com/whirli/whirli-client/pull/187))
 
 ## [0.0.18] - 2020-06-03
 ### 🚀 Features